Batch a whole buffer to write instead of writing tiny bits at a time
Previously, we would flush a write to disk everytime we finish processing
a bulk get, which in almost all cases with small data sizes would not fill
up the buffer completely.

Now, we keep filling up the buffer with multiple bulk gets until it is complete
before we process it and write it to the disk. This should give us quite some
gains in performance.

TODO:
- Clean up code