Support Key filtering via Ketama hashing.
This patch introduces 2 new optional arguments:
'--all_ips_filepath' and '--dest_ips_filepath'
They must either both be provided together or neither should be
provided.

These must point to files that contain all IPs and destination IPs
respectively. The files must be of the format:
   <IP1>:<port1>
   <IP2>:<port2>
   <IP3...

The IP:Port pairs must point to a unique memcached service.
This information is used to filter keys to a subset of destination
IP addresses.

The class KeyFilter exposes a FilterKey() API that returns 'true'
or 'false' depending on whether that key must be ommitted or
dumped respectively.

This 'KeyFilter' class is used in 'ProcessMetabufTask' to decide
whether to track a key for dumping or not.

TODO: Add a unit testing platform and introduce key filter tests.
TODO2: Consider moving all startup options into a YAML config file.