EVCACHE-470: Infinite loop in KeyValueWriter if there's a mismatch in tracking pending keys
We noticed `n_pending_keys_` as 1 but the mcdata_entries_pending_ and
mcdata_entries_processing_ maps were empty.

This causes KeyValueWriter::Finalize() to run into an infinite loop.

This patch uses an assert() statement to crash in this scenario. We'll
use the core file to find out more information on the bug.