[Resume mode] should not resume if key dump from previous run incomplete
The dumper starts dumping data as soon as the first key file is
produced. So the existence of data files does not indicate that the
key dump was complete.

This patch explicitly checks if the key dump was complete by validating
the existence of a checkpoint file called ALL_KEYFILES_DUMPED. If the
dumper is configured in resume mode and this file does not exist, this
patch will clear all the output directories and files from the previous
run and start from scratch.