Fix typo in KeyFilter that caused incorrect keys to be filtered
KeyFilter::FilterKey() checks if a given key hashes to a host that
we have indicated as part of 'dest_ips'.

However, due to a typo, it was filtering out keys that we want but
keeping keys that we actually don't want. This patch fixes that.

NOTE: This is exactly why we need a test suite. TODO soon.