Maintain metrics in a single class and standardize metrics output to JSON
This patch moves metrics that were scattered in different places
to the DumpMetrics class.

This also standardizes the metrics output via a new function called:
DumpMetrics::MetricsAsJsonString() which returns a pretty-formatted
string with the metrics.

This patch was done to enable persisting metrics, which will be a follow
on patch.